src/requirements.txt
google-cloud==0.27.0
#picamera==1.13
Flask==0.12
urllib3==1.22
PySocks==1.6.
RPi.GPIO==0.6.3
